backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
anthropic==0.28.1
google-generativeai==0.5.4
mistralai==0.4.2
openai==1.35.3
qdrant-client==1.9.1
requests==2.32.3
python-docx==1.1.0
pypdf==4.2.0
sentence-transformers==3.0.0
torch==2.3.1
